requirements.txt
pillow
thinker